Post-poll-redesign cleanup — delete orphaned files + dead UI + drop poll_note
Mechanical cleanup of code left flagged for removal after the date-poll
stacked-cards redesign (PR #286) landed.

Deleted:
- DatePickerPanel.tsx — zero callers. DatesSheet was migrated to the
  inline FullRangeCalendar.
- DatePollGrid.tsx — superseded by DatePollStackedCards. Hoisted the
  PollMember / PollWindow / PollWindowVote / VoteAnswer types into a
  new datePollTypes.ts and repointed the two import sites
  (DatePollStackedCards + DatePollCard) before deletion. (The unrelated
  marketing DatePollGrid in HeroSection.tsx is a name collision — left
  alone.)
- ItineraryIntroModal.tsx — zero callers. Stage model dropped earlier.
- ItineraryPanel _disableItinerary mutation hook + the void-suppressed
  useState import. Server procedure (trips.disableItinerary) kept on
  standby per the orienting comment, but the void'd client hook is gone.
- DatePollCard resetPoll mutation hook + the two void-suppressed
  references. Server procedure (datePoll.resetPoll) kept; tests still
  cover it.
- MemberEditor's `void Mail;` lint-suppressor — Mail is used in the
  body for a real reason, the void was a leftover.
- FreshTripGuide `onOpenDatesSheet` prop — dead pass-through with
  nothing reading it inside the guide. ItineraryPanel still uses the
  prop locally (DismissedEmptyState handoff), just stops passing it
  down.

Server / DB:
- datePoll.updatePollNote procedure deleted. The Crew Instructions
  surface that consumed it was removed in PR #286.
- date_polls.poll_note column dropped via new migration 021. Reads
  and writes in datePoll.get / addWindow's optimistic cache shape
  also cleaned up.

Type-check + lint clean. datePoll.test.ts (21 tests) passes.
